Added scroll wheel support.

+++ selst.h
@@ -1,20 +1,20 @@
 #ifndef SELST_H
 #define SELST_H
 
-/*  selection endpoint types.
- */
+#include <stdint.h>
+
+//  selection endpoint types.
 enum seltype {
 	SCREENSEL,
 	SAVEDSEL,
 	NOSEL
 };
 
-/*  structure describing a selection endpoint.
- */
+//  structure describing a selection endpoint.
 struct selst {
 	enum seltype se_type;
-	int se_index;	/* index into the sline or screen array */
-	int se_col;	/* column of the character */
+	int16_t se_index;	// index into the sline or screen array
+	uint8_t se_col;		// column of the character
 };
 
 #endif//!SELST_H
